refactor: Rename data files with more appropriate names
