+++ src/components/organisms/TastingCategory.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { Header, Grid } from 'semantic-ui-react';
 import TastingItems, { Attributes } from 'containers/molecules/TastingItems';
-import { CategoryTitles, SubCategoryItems } from 'data/tastingCategories';
+import { CategoryTitles, SubCategoryItems } from 'data/tastingSheet';
 
 type Props = {
   categoryTitle: CategoryTitles;
